requirements.txt
googletrans==4.0.0rc1
python-dotenv
google-generativeai
termcolor